# OpenAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "F1 API",
    "description": "An easier way to use the [ergast.com](https://ergast.com/mrd/) F1 API, with correct types!",
    "license": {
      "name": "MIT",
      "url": "https://github.com/niek-o/F1API/blob/main/LICENSE.md"
    },
    "version": "1.5.0",
    "x-logo": {
      "url": "https://upload.wikimedia.org/wikipedia/commons/f/f2/New_era_F1_logo.png"
    }
  },
  "paths": {
    "/calendar/{season}": {
      "get": {
        "tags": [
          "Season"
        ],
        "summary": "Get Calendar By Season",
        "operationId": "get_calendar_by_season",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "integer"
            },
            "name": "season",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                },
                "example": {
                  "events": [
                    {
                      "season": 2022,
                      "round": 1,
                      "url": "https://://en.wikipedia.org/wiki/2022_Bahrain_Grand_Prix",
                      "raceName": "Bahrain Grand Prix",
                      "Circuit": {
                        "circuitId": "bahrain",
                        "url": "https://://en.wikipedia.org/wiki/Bahrain_International_Circuit",
                        "circuitName": "Bahrain International Circuit",
                        "Location": {
                          "lat": 26.0325,
                          "long": 50.5106,
                          "locality": "Sakhir",
                          "country": "Bahrain"
                        }
                      },
                      "date": "2022-03-20",
                      "time": "15:00:00Z",
                      "FirstPractice": {
                        "date": "2022-03-18",
                        "time": "12:00:00Z"
                      },
                      "SecondPractice": {
                        "date": "2022-03-18",
                        "time": "15:00:00Z"
                      },
                      "ThirdPractice": {
                        "date": "2022-03-19",
                        "time": "12:00:00Z"
                      },
                      "Qualifying": {
                        "date": "2022-03-19",
                        "time": "15:00:00Z"
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Calendar not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/event/next": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get Next Race",
        "operationId": "get_next_race",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextEvent"
                },
                "example": {
                  "season": 2022,
                  "round": 17
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/event/previous": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get Previous Race",
        "operationId": "get_previous_race",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextEvent"
                },
                "example": {
                  "season": 2022,
                  "round": 17
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
    "/event/{season}/{round}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get Event Details",
        "operationId": "get_event_details",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Season",
              "type": "integer"
            },
            "name": "season",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "title": "Round",
              "type": "integer"
            },
            "name": "round",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                },
                "example": {
                  "season": 2022,
                  "round": 1,
                  "url": "https://://en.wikipedia.org/wiki/2022_Bahrain_Grand_Prix",
                  "raceName": "Bahrain Grand Prix",
                  "Circuit": {
                    "circuitId": "bahrain",
                    "url": "https://://en.wikipedia.org/wiki/Bahrain_International_Circuit",
                    "circuitName": "Bahrain International Circuit",
                    "Location": {
                      "lat": 26.0325,
                      "long": 50.5106,
                      "locality": "Sakhir",
                      "country": "Bahrain"
                    }
                  },
                  "date": "2022-03-20",
                  "time": "15:00:00Z",
                  "FirstPractice": {
                    "date": "2022-03-18",
                    "time": "12:00:00Z"
                  },
                  "SecondPractice": {
                    "date": "2022-03-18",
                    "time": "15:00:00Z"
                  },
                  "ThirdPractice": {
                    "date": "2022-03-19",
                    "time": "12:00:00Z"
                  },
                  "Qualifying": {
                    "date": "2022-03-19",
                    "time": "15:00:00Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Event not found"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                },
                "example": {
                  "message": "Service unavailable"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Calendar": {
        "title": "Calendar",
        "required": [
          "events"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "events": {
            "title": "Events",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          }
        }
      },
      "Circuit": {
        "title": "Circuit",
        "required": [
          "circuitId",
          "url",
          "circuitName",
          "Location"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "circuitId": {
            "title": "Circuitid",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "circuitName": {
            "title": "Circuitname",
            "type": "string"
          },
          "Location": {
            "$ref": "#/components/schemas/Location"
          }
        }
      },
...
      "Event": {
        "title": "Event",
        "required": [
          "season",
          "round",
          "url",
          "raceName",
          "Circuit",
          "date"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "season": {
            "title": "Season",
            "type": "integer"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "raceName": {
            "title": "Racename",
            "type": "string"
          },
          "Circuit": {
            "$ref": "#/components/schemas/Circuit"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "FirstPractice": {
            "$ref": "#/components/schemas/Session"
          },
          "SecondPractice": {
            "$ref": "#/components/schemas/Session"
          },
          "ThirdPractice": {
            "$ref": "#/components/schemas/Session"
          },
          "Qualifying": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Location": {
        "title": "Location",
        "required": [
          "lat",
          "long",
          "locality",
          "country"
        ],
        "type": "object",
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "long": {
            "title": "Long",
            "type": "number"
          },
          "locality": {
            "title": "Locality",
            "type": "string"
          },
          "country": {
            "title": "Country",
            "type": "string"
          }
        }
      },
      "Message": {
        "title": "Message",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        }
      },
      "NextEvent": {
        "title": "NextEvent",
        "required": [
          "season",
          "round"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "title": "Timestamp",
            "type": "number"
          },
          "season": {
            "title": "Season",
            "type": "integer"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          }
        }
      },
...
      "Session": {
        "title": "Session",
        "type": "object",
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "string"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}